Move encode and decode macros to separate files
